MXS-4596 Do not remove +/- signs before converted values
This is vaguely related to the Jira. For consistency drop code
that tries to avoid only removing a '-' prefix to a number.
There is not enough context for the code to reliably do it,
and for consistency, it would have to do the same for all
operators. So now:
Sql  Canonical
-1   -?
1-1  ?-?
a-1  a-?
1-a  1-?