MXS-4599 Minor cleanup
Do not use static in __m256i bitmask_lookup(), the init is
a few asm instructions only. Also, the init value 128 was changed
to char(128) to keep some tools from nagging.

There was one more signed value where unsigned (or no cast)
should have been used, as well as a logical "or" where a bitwise
"or" was the proper operator.