Only check stale connections if needed
Each read iterates through all of the available backends in
readwritesplit and the status and the rank are extracted there. By
detecting stale connections in this function, the cost of checking can
be eliminated in the very common case of there being no stale
connections that need to be discarded.

Also changed get_backend_priority() to use the status value that was read
earlier. This makes sure it's read only once and that it remains
consistent across all operations.