De-virtualize mxs::Target::status()
A change in server status happens rarely and recursively calculating the
status every time it is requested is a relatively heavy
operation. Calculating it only when it changes makes reading it a faster
operation. In addition, de-virtualizing the function allows the value to
be inlined by the compiler.

There's room for optimization in how the per-server status updates
propagate to the services. These could be bundled with the GTID messages
for the MainWorker.