Use EPollTick for interval timers
A lot of the time went into calling the more accurate mxb::Clock::now()
when the granularity of the EPollTick timer was good enough.

Renamed the IntervalTimer class into IntervalTimerBase and implemented
the clock type selection via templates. Added a type declaration for the
IntervalTimer that now uses mxb::NowType::EPollTick as the clock source.