Simplify log bitmask checks
The extra logging bits that need to be enabled are known when the
current DCB being processed is set. By copying the session's bitmask
into the thread-local extra bitmask variable, the cost of individual log
messages is reduced.

Given the fact that there are dozens of log calls per request-response
cycle, the cost of the check started to show up in profiling. The old
version took 0.6% of the total CPU time.