Add mxb::small_vector
The small_vector is a generalization of the std::string optimization for
small strings. The most known implementations of these types of
containers are in LLVM and Boost. The container is optimized for use as
a small sort buffer that does not allocate heap storage if the number of
elements is below a compile-time constant.

For MaxScale, this is beneficial for readwritesplit that used a
std::vector to sort a handful of pointers. By swapping out the
std::vector for a mxb::small_vector, the number of heap allocations
drops down to zero if the number of backend servers remains below 4. The
choice of how many values to store internally was mostly arbitrary but
with 4 values the size of the Candidates type ends up being exactly 64
bytes.

Added a unit test that makes sure the behavior of the container is
identical to a std::vector for the supported operations.