Add Parser::is_multi_stmt()
The QueryClassifier class uses the information to detect when queries
need to be routed to the current master server. For every query that
gets executed, one call to maxsimd::is_multi_stmt() was made. By storing
the result in the query classifier cache, the cost of this is eliminated
for all cached entries.