Move cache exclusion checks into the parsers
This removes the redundant computations that were done for every call to
the mxs::Parser API functions. Moving the cacheability information into
the base class also simplifies the check by moving the data closer to
where it's used.

The m_canonical was also redundant as the information was never used
outside of debug assertions. It can be replaced with a reserved value
that is stored in m_info_size_before that tells when to insert the
value.